1) Add new question & answer 2) Run model, and log request, response - all goes well (in future - wants try all previous questions) 3) Also when calculate total estimation, output float value

+++ src/main/kotlin/Main.kt
@@ -56,6 +56,6 @@ fun main() {
     }
     println(responses)
     println(estimations)
-    val totalEstimation = estimations.sum() / estimations.size
+    val totalEstimation = estimations.sum().toFloat() / estimations.size
     println(totalEstimation)
 }
